requirements.txt
speedtest-cli
mysql-connector-python
requests
PyYAML